app/backend/requirements.txt
boto3
flask
python-dotenv
flask_cors